Fix print-text-format
The bool special case we are removing has
been deprecated.
We also modify the type and signature of the function
print-text-format. It's odd to do a destructuring-bind on keywords,
so instead make the bound keywords be the parameters for the
function.
The function need not be a method so make it a function.
Finally override the print-object function for proto-messages
to use print-text-format

PiperOrigin-RevId: 311647204