We use proto-parent field on protobufs to make the qalified-name and to find the schema.
I don't (yet) have a good answer for getting
the schema, but it we never try to find it
and we remove proto-parent in production so that's never
a problem.
For making the qualified name, there's no reason
to do the recursive qualified name work,
just look at the parents qual name and tack
on your name.

PiperOrigin-RevId: 314038268