The concept of proto-parent is understandable but not needed and the concept of real-package-name is absurd.
Remove proto-parent, the only places it's currently used
will either
a) Cause a failure as we never set the field in the constructors
b) We've removed it in favor of setting a service name
on a protobuf-method.

We get the package for a message from the methods class
symbol. If it's not there we have no hope in
getting the field anyway. But the way thats used is
equally absurd and should probably just be removed...

PiperOrigin-RevId: 314127460