Process files in GenerateAll so that imports are first
While CodeGeneratorRequest.proto_file is in topological order, the input to GenerateAll is in the same order as CodeGeneratorRequest.files_to_generate. That's the same order that the .proto files appear on the command-line, which is not required to be in a particular order. So if we're processing multiple files and one imports the other, we need to generate code for the imported one first.

PiperOrigin-RevId: 314517236