Move *protobuf[-rpc]-package* to the parser
...and no longer use it in the define-* macros. It's sufficient to
define things in the current package. Moved them to the parser rather
than deleting them altogether because the parser may be deleted and
I don't want to spend time on it.

Also fixed a few (intern (format nil "~A"))s to call nstring-upcase
since ~A respects *print-case* and this leads to bugs.

PiperOrigin-RevId: 314532802