Add copy of proto2 extensions to google/protobuf/
This is a quick fix to the currently broken makefile. Since
cl-protobufs now supplies descriptor.proto, the import statement
of proto2-descriptor-extensions must reference that local version.
However, this is inconsistent with how we build the protoc plugin,
and thus the makefile breaks. Adding this copy with the original
import statement allows the makefile to work.