Fix imports-test so that imports are called in the correct order
This test was originally intended to check if we could load .proto
files out of order with respect to their imports in a .asd build
rule. This currently doesn't work that way, and for the future we
will require that users load their .proto files in the correct
order.

Now, import-test only tests the functionality of loading .proto
files that have imports.