Fix define extends.
The weak hash table is currently set to remove
it's entry if it's value becomes inaccessible.
But we have a test that tried

(let ((a (make-base)))
  (setf (extension a) (make-extension)))

Simply, as soon as the extension is set it's
only accessible through the hash table,
this it's available for garbage collection
and the slot is removed.

Instead we should set it to be keyed on the key,
that is the object, so the entry is only removed
when it's either
a) deleted
b) cleared

PiperOrigin-RevId: 316301258