Let serialize.lisp be loadable in any any common lisp (or at least ccl)
Serialize.lisp uses keywordicate, a silly sb-int function
that makes keywords.
But why?
One thought is to call alexandria make-keyword but that seems
silly when we only want 1 function.
Instead just make the keyword the old-fashioned way.

We also featurise fast-utf8 to fall back to
babel if we don't have sbcl

This leave buffers.lisp and wire-formats.lisp

PiperOrigin-RevId: 316371778