Add map-of type
This allows for easy access to the key & value types defined
in a map without the need for a separate model-class object.
Unsatisfyingly, we still define the "map entry" message in protoc
and use that as a sort of pseudo-map-descriptor, as it contains
lots of useful things like (de)serializing subroutines and
initforms. As Jon has pointed out, using these serialization
subroutines is going to be far too slow, and it'll be better to
just rewrite them. At that point, it'll probably be easier to
just duplicate the default-form generating code too, so we can
remove that message definition and rely only on the proto:map-of
type.