Add and generate special map forms
The "common" forms generated for map types are unsafe. They
expose the entire map to the user so two things happen:
1. Type checking is turned off.
2. We have no way to see if the map is set or not.
This is necessary, as users will want access to the map, but we
also need some kind of safe accessor function. So, this commit
adds the commands "get, put, remove" which do what you would think
they do on a hashmap. These are type checked, and update the
%%is-set flag properly.