Add compatibility code to cl-protobufs to compile in other lisps
cl-protobufs should be compile-able in multiple common-lisps,
not just sbcl. This makes it compileable in at least
CCL and ABCL, and wire formats tests run in both.
It still may not work properly in languages other then sbcl,
but this is a step.

We do a few things:
We featurize buffer-sap in buffers.lisp to be only SBCL
We remove two include forms (type and API) from cl-protobufs.asd
as we get some strange conflict in ccl.
We set the in-package line in float-bits that seemed to have
been missed.
Add the compatability code in wire-format
Replace make-package with defpackage ... (:use) in file.cc
because for some reason ccl includes :ccl in your use statament
by default...
Remove a test for ccl and abcl (different tests) because they
don't do the proper cut-down for bits in uints.

PiperOrigin-RevId: 317676642