Add proto-type->keyword hack
This is used whenever we need to extract type information from a
map type, since some places use types and others use keywords to
reference the same thing (e.g. :int32 vs proto:int32). The set-type
of maps is always the type, so proto:int32, but the serialization
code uses keywords. This function also contains a bad hack where
we rely on type being of the form (or null [type]) when type is not
a primitive.