Add additional map tests
Currently it seems that proto-equal doesn't work properly with
maps, as the tests fail for enums (even though the maps are
indeed equivalent). This seems strange since equalp is supposed
to work for hash-tables. So, the tests fail, but that seems to
be a problem with the tests rather than serialization.