Remove strange handling of (or null ...) types
Previously, (or null ...) was always assumed to be a message type,
and we would use the structure of a type as a condition to check
if the type defined a message. Now, protoc will only output the
name of the message itself, and we create the (or null ...) wrapper
when needed (during the creation of accessor/clear functions).