Update map accessor functions
This is done in an attempt to closely mirror the accessor functions
of vanilla lisp hash-tables. Now, message.field-get has the same
parameter signature as gethash (and similarly for message.field-remove
and remhash). Additionally, to add to a hashmap, there is a setf
syntax: (setf (message.field-get [key] [msg]) [val]).