Rename enum-value-descriptor's value slot to name
The "value" of an enum is the numeric value. The "name" of an enum key/value pair is the textual key corresponding to a value. This CL changes the value slot to name since it holds a keyword symbol. The next CL will rename index to value. This puts us more in line with the C++ protobuf terminology.

PiperOrigin-RevId: 320002090