Replace keywordp checks with "scalarp"
Since the class of a field with a scalar type is always a keyword,
calling keywordp on the class will tell you if the type is a
scalar. However, this is totally opaque and non-obvious (and was
also false when maps briefly had their class set to :map). This
fixes that by introducing the scalarp defun, which checks if a type
is a member of the list of scalar types.