Fix some compiler notes.
This is a minor foray into all the compiler notes we're
seeing. Many of them indicate failures to optimize the code. This
takes us from 72 to 41 such notes in wire-format.lisp.

Hint: sbcl doesn't like when you modify a function parameter.

Made a new function, make-wire-tag, to call when you know the exact
wire type to encode. make-tag continues to be used for translating a
protobuf "class" (:int32, etc) to a wire type.

PiperOrigin-RevId: 320680195