Add oneof structs
Oneofs consists of two structs:
The first is the oneof-descriptor which holds all static data
relevant to a oneof. This includes name/slot info and a vector
which contains all of the field descriptors for the fields defined
inside of the oneof.

The second struct is the struct that is stored on a message struct
whenever it has a oneof. This struct has two slots: value and
set-field. Value stores the data of whatever field is currently
set in the oneof. set-field holds an identifier that allows us to
quickly identify which field in the oneof is currently set (so that
we can extract information like the field descriptor, type, etc.)