Add function names for oneof accessors
Fields defined in a oneof keep their normal accessors, and using
those will result in the correct behavior for the other fields in
the oneof. However, there are three special functions created for a
oneof.

For a oneof 'oneof' in a message 'msg', we create 'msg.has-oneof'
, 'msg.clear-oneof', and 'msg.oneof-case'. The first two do what
you would expect. The first returns whether or not any field in the
oneof is set, and the second clears any field in the oneof. The last
function returns a keyword which identifies the field in the oneof
that is set. So, if my oneof contains a field such as
'int32 intval = 1', and 'obj' has 'intval' set, then the call
'(msg.oneof-case obj)' will return ':intval'. The keyword is just
the name of the lisp-name of the field.