Fix generic deserializer not handling multiple oneof fields
The generic deserializer first parses in all fields and sorts their
data by field number. Afterwards, the object is created. This
cannot work for oneofs, as oneof spec says that if two fields
in the same oneof appear on the wire, then the last one is set.
This is a problem, since the aforementioned sort by field number
destroys all information about the fields order on the wire.

The solution is to simply ignore this sort for fields which are in
a oneof, and always append them to the end of list of arguments.