Remove check that enum is not default value.
This check doesn't make much sense. Why do we not want to parse
an enum field if it has the default value? Is this somehow supposed
to save space? Besides, we don't do this for repeated fields, so
why would we do it here? Removing this as we want to remove any
dependency on field-descriptor from this function.