Remove remaining test failures for ccl
In quicktests we make test-field-maps only run
for sbcl. I'm not sure why it's failing on SBCL
so at some point we should take a closer look.
Its not quite a protobuf functionality test though.

In zigzag test I remove an assert that depends
on a specific name. The name is different in
different common-lisps so make it
only be evaluated on sbcl.

PiperOrigin-RevId: 323415258