Fix serialization-tests for CCL.
This cl fixes serialization-tests, and
after this cl only zigzag-test and quick-test are failing
in ccl.
zigzag is due to a name difference (uint64 vs unsigned-byte 64)
and quick-test has something to do with hash maps.

The text format change removes the usage of make-instance
for everything but SBCL. Make-instance has no gaurantee of
working, but we know it does for SBCL, thanks @dougk.

Wire-format is changed to check the number of bytes can be written
to buffer and then just use fast-octets-out.
For UINTs this means finding the length first.
For everything else it's just a speedup.

Finally for non-sbcl we don't generate the special
deserialization function, instead it's a lambda
attached to the symbol plist.
So make a block and return from the block at the top
of the lambda.

PiperOrigin-RevId: 323423578