use buffer-ensure-space and fast-octet-out instead of (setf index)
Currently we manually set the bytes into BUFFER.
We do this by checking if we should use scratchpad
or the current octet-buffer-block and manually indexing
in with index.
This is silly.
Instead just call buffer-ensure-space
which gaurantees that there is enough space in
the buffer, and then freely call fast-octet-out.

PiperOrigin-RevId: 323437995