Remove IDX binding from process-field
In the context of DEFINE-GROUP and DEFINE-MESSAGE, there's no
reason to hang on to the field number of the most recently
processed field. Any error about overlapping field numbers is
caught by protoc. However, for DEFINE-EXTEND, we still need to
hold on to the index to check if the index has been filled yet.