Check for lazy bytes at each recursive call of generic serializer
This does a few things:
1. Fixes an uncaught bug in generic serialization. The function
emit-non-repeated-field did not check for lazy bytes. So, if a field had
a %bytes field, then the serializer would ignore it and serialize
whatever was stored (which is always the default value). Now, since the
serialize-object function checks for a %bytes field, this bug has been
fixed.
2. Cuts down on repeated code. Now we don't need to check for the
existenceof these bytes in emit-repeated-field AND
emit-non-repeated-field.
3. Allows serialize-object to efficiently serialize objects which have
lazy bytes. Previously, only serialize-object-to-bytes could do this
(which in turn calls serialize-object). So, if one calls
serialize-object-to-bytes there is one redundant check, but this only
happens once at the top level of the function so it's most likely fine.