Fix broken test
This assertion is supposed to check that the following behavior works:
If we have some bytes which are a serialized lazy proto, then we can
deserialize it lazily, and then serialize it and get the same thing
back. However, in this test, there were calls to the lazy field's
accessors in between the deserialization and serialization steps. So,
the lazy fields were being deserialized at that point, which should not
be part of the test. This change fixes that by binding RESTORED to a new
deserialized proto, and then directly serializing that.

This test now catches the error that is fixed by a previous commit.