Prioritize lazy serialization over custom-serializer calls
There are some messages in an application of cl-protobufs with the
following situation
1. They have a bound %BYTES slot
2. They have a custom serializer
3. The custom serializer does not check for %BYTES before attempting a
serialization.
In this case, the serializer code would first check for a custom
serializer, find it, and then call that. This ignored the %BYTES slot,
and thus nothing would be serialized. This commit ensures that we always
check for %BYTES first.