Ensure that singular fields do not have a bit in the is-set vector
This was duplicated data. This bit only stored the answer to the
question "is this currently set to the default value"? I figured this
would be helpful as I thought that computing the answer to that question
was expensive, how it turns out to be pretty cheap. So, it's OK to do it
in the has-* function.