Update serialize.lisp to handle singular fields
Since singular fields no longer have an is-set bit, the generic
serializer needs special handling for singular fields. Note: There is
code duplication between this, text-format.lisp, and the has-* function.
This can be fixed by calling the has-* function here, but this requires
a lookup in a hashtable of functions, which is likely too slow for a
serialization function.