Verify that the report file was written and the tests ran successfully
We currently only check that the report file has no lines saying
"not ok". This is required but not sufficient.
First check the report file exists.
Then verify that there is no line saying "not ok"
and some line saying "ok".
With that we can be sure some tests passed.

PiperOrigin-RevId: 325026321