Minor updates to get ABCL closer to working
In serialize.lisp we remove the map-direct-slot-p
because both ccl and ABCL are getting confused by it.
It's not pulling much weight anyway.
We also add a nil check as one of the function
could be indexing into a nil.

In wire-tests we remove a fixnump requirement.

In wire-tests.lisp we update what ccl can handle.

Does not fix all bugs but almost all of the
non-oneof ones...

PiperOrigin-RevId: 325470434