Update field-descriptor slots.
TYPE is the base lisp type of the field. "Base" here means that if a
field is repeated then we put the underlying type rather than (list-of
type).
CLASS is a keyword indicating if the type names a scalar, enum, message,
group, or map. For scalars this information is redundant, and at
run-time it is redundant since you can just call (find-* ...) on the
field's type. However, we need this information at compile time, and the
run-time work-around is a bit of a hack.
CONTAINER indicates which lisp container is used for repeated fields. It
is either :list, :vector, or nil (non-repeated types only). This data
could probably be stored in the LABEL slot, so :repeated would be
replaced with :list or :vector. However, I'm not so sure about that
since that seems to muddy the waters between proto spec and our own
lisp-only extensions.