Fix bug in non-repeated-bool-field defun.
This function previously returned true if and only if the field form
contained 'cl:boolean' AND contained ':repeated'. However, field forms
do not contain ':repeated', and instead they contain (:repated :list) or
(:repeated :vector). The reason this function worked, is because
repeated boolean fields would have the type (list-of cl:boolean), and
hence the form would not contain 'cl:boolean. This fixes that bug.