Add more general handling for 'null' fields.
'null' can also appear as the value for a non-repeated field. This value
represents the default value of the type. Since JSON mappings are only
documented for proto3 semantics, this means that we should simply not
set the field to anything in this case. In proto2, this means that the
has-* function should return NIL if a field is parsed with a 'null'
value. This commit adds a test for this behavior.