Update oneof comment to add clarity
This comment was written this way as there was a piece of code which
used the truth value of the field-offset slot as a way to decide if a
field was an extension or not. Since oneof field descriptors have no
offset, this would make oneof fields interpreted as extension fields.
This has since been fixed, and is no longer a worry.