Add the field 'container' to field-descriptor
Currently in cl-protobufs we set the 'default' field
of the field-descriptor to whatever is said in the lisp schema
file. Then if we want to know if it's
a) repeated
b) vector or list
we check the field to determine if the second value is list
or vector. Instead we should just make a container field.

This is part of
https://github.com/qitab/cl-protobufs/pull/156

PiperOrigin-RevId: 330771256