Add proto-kind and remove proto-message-type in field-descriptor
In the field-descriptor class we have a field called
'message-type'. The field need not be a message,
so it should definitely not be there. Instead
we need a field that says what 'kind' ie
:scalar, :enum, :message,... is actually in the field.
We use this data to compute defaults of the field
and other such requirements.

We update field.cc to give us this data.

This is work to get us ever closer to Bens
pull request:
https://github.com/qitab/cl-protobufs/pull/156

PiperOrigin-RevId: 330795670