Have type in the cl-protobufs field descriptor be the base type of the field
We already have the list type (vector or list) in the container field,
and we already have whether it's repeated or not in the label field,
so we don't need to store the extra data in the type field.
Instead use type to determine the base type of the field.

This is more of:
https://github.com/qitab/cl-protobufs/pull/156

PiperOrigin-RevId: 330997074