Add typing to protobuf make-vector in get-default-form
Currently we make a general re-sizable vector in that
we add by default into the protobuf object
when a user request a vector repeated object.
Thanks to Bens work we know have the type information
so we can make a re-sizable vector with
the correct element type.
We should add a fill point so vector-push-extend
can be used to add into the vector.

PiperOrigin-RevId: 331681819