Add a push function to add elements to a repeated protobuf object
We have the ability to use vectors instead of lists as the container
type for a repeated field in cl-protobufs.
We can use vector-push-extend to add to this vector field
but that wont set the has bit in the bit vector.
This is not a problem for lists because list uses
a push macro that inherently uses setf.

We also add tests for this new functionality.

PiperOrigin-RevId: 331911463