Make the has function for repeated fields check for emptiness of there containers.
Currently the has function for a repeated field in common lisp uses the
is-set bit vector. This is wasteful, why do we need to store this information
in a bit vector when all we need is to check the size of the repeated
container. It gets worse, if someone setf's the field to an empty
vector we think it is set!
Here we either
a) Check if it's length is zero for a vector.
b) Check if it's nil for a list, empty == nil.

PiperOrigin-RevId: 331930183