Don't :use packages in test code
I started work on a different PR and ran into name conflicts with the test code because it uses :use to import cl-protobufs and the generated code for test protos.

Let's not do that. Besides potential for more conflicts, it reduces the overall clarity of the code in my opinion. This PR removes :use from several tests and imports the test protos with package local nickname "pb", for brevity. This is short, but it should be clear if the same pattern is followed in all the tests. I'll convert more as time permits.

Also:
* Renamed deserialize-object-to-bytes-test.lisp to deserialize-test.lisp
  ("to" was inaccurate and the name is unnecessarily long).
* Renamed custom-methods.lisp to custom-methods-test.lisp (standard)
* Use clunit2 assertion macros, to improve test failure messages.
PiperOrigin-RevId: 332324153