Simplify define-proto.lisp and remove clos-transforms.lisp
With the recent change the the message and field descriptor
classes we no longer need to keep transforming for proto
types to lisp types and back. This cl takes some of the code in
clos-transform and adds it to utilities, but mostly deletes
the code in clos-transform.

We can do this because we store the actual type of
an object and not the list-of type or the "proto type".

PiperOrigin-RevId: 339985919