Move api.lisp to message-api.lisp
We try to compile both api.lisp and the well
known type proto api.proto.
This is all fine for bazel, but for ASDF
this causes a conflict in the fasl files
which both want to be called api.fasl.
When we run the asdf build command once
it doesn't matter, but ASDF tries to save
these fasl files.
The next time we try to load cl-protobufs ASDF says:
"Hey look, an api.fasl file!"
Sadly, at the top of this file is an validate
import statement for the type.proto file, and as
we are trying to load the elements in api.lisp and
not api.proto we don't yet have type.proto.

This is easily fixed with a rename.

PiperOrigin-RevId: 341475519