Change feature flags for float-bits and wire-format
In cl-protobufs we need to know how to work with 64 bit
floats. We have a file for float bits (which needs massive work),
Currently we don't load it for SBCL as SBCL has float functionality
and who would use a non-64 bit OS? Well, raspberry pi...
So enable float-bits for non-64 bit SBCL instances.
Also, fix the feature flagging for wire-format (which again needs
work...)

PiperOrigin-RevId: 345296794