Specify utc-time when printing a timestamp in cl-protobufs.
In cl-protobufs we omit the timezone when telling local-time
to format-timestring. Previously this worked because
we were calling dolist on a vector which failed. This was wrapped in
a handler case in a cond which returned nil. The
second (default) possibility in the case was to use UTC time
so local-time:*default-timezone* was always UTC.
We removed the dolist so we no longer fail and get the
correct default-timezone as specified in /etc/localtime
which is not what we want... Hence the fix.

PiperOrigin-RevId: 350155459