Change enum-valued slot type specifiers to 'keyword
Tests failed in non-optimized settings. Enum-valued slots must be typed
as just 'keyword so that we can store :%undefined-n in them.

The diff doesn't clearly show the effect of the change. Expanding the
scope of the rebinding of `type` causes it to be used in the binding
of `field` as well.  (The cond changes are just cosmetic.)

PiperOrigin-RevId: 353312965