Have cl-protobufs print and parse text-format without type
Currently, cl-protobufs prints:

ProtoTypeName {
...
}

when print-text-format is called. This is not
in the text-format spec, so it should not be printed.
Similiarly, we remove the bit of code that checks
for ProtoTypeName when parsing a message
in text-format

It's still not a proper text-format printer/parser
but this makes it closer...

PiperOrigin-RevId: 380073122