defmethod print-object implementation
Implement a method defined for objects of class `message`. When an instance of `message` class is printed using the `print` function, this method will be invoked.

The purpose is to customize the textual representation of the `message` object when it's printed to a stream.

It invokes `print-text-format`, sending `message` object, the output `stream` and null`:pretty-print-p`.

Create a test case that asserts defmethod `print-object` by matching a formatted string representation of the parsed message msg to an expected literal string

PiperOrigin-RevId: 542866475