Work on cl-protobufs support for feature presence in editions.
Proto editions is here, and we wish to support it.
One of the things to support it is to add
support for setting presence to implicit or
explicit.

For backwards compatibility we set the field presences
for maps and repeated fields to explicit.
This doesn't produce an interesting has function,
just a has function that checks if the type
is empty, but we use it in tests.

PiperOrigin-RevId: 699362806