Further optimize overloaded field accessors for SBCL
If N different type-specific accessors are all dispatched to by a certain overloaded name, but in each object type to which the overload pertains, the same field index is accessed, then the overloaded function can access the object at a known slot index without using a function call. Or if there are exactly 2 different possible field indices accessed by the overloaded name, it can be a single IF rather than a call. For more than 2 choices, it will do what it did prior to this enhancement.

PiperOrigin-RevId: 778710113